# Pool file handles used by StreamBinaryReader components