… CHANGELOG.md
… cmd/mimir/config-descriptor.json
… docs/sources/mimir/operators-guide/configure/about-versioning.md
… docs/sources/mimir/operators-guide/configure/reference-configuration-parameters/index.md
… pkg/storegateway/indexheader/encoding/encoding_test.go
… pkg/storegateway/indexheader/encoding/factory_test.go
… pkg/storegateway/indexheader/encoding/reader_test.go
… pkg/storegateway/indexheader/header.go
… pkg/storegateway/indexheader/header_test.go
… pkg/storegateway/indexheader/index/postings.go
… pkg/storegateway/indexheader/index/symbols.go
… pkg/storegateway/indexheader/index/symbols_test.go
… pkg/storegateway/indexheader/lazy_binary_reader_test.go
… pkg/storegateway/indexheader/reader_pool.go
… pkg/storegateway/indexheader/stream_binary_reader.go
… pkg/storegateway/indexheader/stream_binary_reader_test.go