@56quarters 2022-12-13
I ended up implementing the [change described here](https://github.com/grafana/mimir/pull/3691#discussion_r1045981910) because the logic for cleanup was getting a little confusing being split between the `FileReader`, `Decbuf`, and `DecbufFactory`. Now callers just call `Decbuf.Close()` (`io.Closer`) and anything that was pooled is put back into whatever pool it came from.